fix(haskell): make Timestamp Ord lawful and Issue position fractional (IRI-45)
Two defects, both where the type disagreed with what it mirrors.

Timestamp had an inconsistent Eq/Ord pair. Parsing is deterministic, so
timestampUtc is a function of timestampText and the derived Eq was byte
equality — while compare looked only at the instant. Two spellings of one
instant therefore gave (a == b, compare a b) == (False, EQ), which violates
the Eq/Ord law and makes Set/Map drop a value equality called distinct.
compare now falls back to the retained wire text, so chronological order
stays primary and compare a b == EQ and a == b are the same statement.
ordRefinesEq exposes the law as a predicate, matching the restrictionRoundTrips
idiom already used in Ubiquity.Actor.

Issue position was Int; it is fractional indexing. Board order is maintained
by writing the midpoint between two neighbours, so the column is FLOAT
(001_init.up.sql:68), float64 on the wire (issue.go:47) and z.number() in the
TS schema (schemas.ts:487). Live data carries -1.5 and the tool died decoding
it. An integral type here denies the operation the field exists to support.

Adds a base-only test-suite for the ubiquity laws; picking a test framework
for the Haskell line stays the VERIFY lane's call.

+++ packages/multica-api/src/Board/Issue.hs
@@ -154,10 +154,22 @@ instance FromJSON IssueLabel where
 
 -- | An issue.
 --
--- Three numeric fields that are not identifiers: 'issueNumber' (composes with
--- the workspace prefix into 'issueIdentifier'), 'issuePosition' (board order,
--- and negative in practice), and 'issueStage' (nullable, and @null@ on the
--- large majority of live issues — its absence is ordinary, not exceptional).
+-- Three numeric fields that are not identifiers, and they do not share a
+-- numeric type: 'issueNumber' (integral; composes with the workspace prefix
+-- into 'issueIdentifier'), 'issueStage' (integral, nullable, and @null@ on the
+-- large majority of live issues — its absence is ordinary, not exceptional),
+-- and 'issuePosition'.
+--
+-- 'issuePosition' is 'Double', not 'Int', and the difference is structural
+-- rather than a matter of range. Board order is maintained by /fractional
+-- indexing/: to move a card between two neighbours the server writes the
+-- midpoint of their positions, so the column is @position FLOAT NOT NULL@
+-- (@001_init.up.sql:68@), @float64@ on the wire (@issue.go:47@) and
+-- @z.number()@ in the TS schema (@schemas.ts:487@). Values are commonly
+-- negative and commonly fractional — a live board yields @-1.5@ — because
+-- density between neighbours is exactly the property the scheme buys. An
+-- integral type here does not merely truncate; it denies the operation the
+-- field exists to support.
 data Issue = Issue
   { issueIdOf :: IssueId
   , issueWorkspaceId :: WorkspaceId
@@ -173,7 +185,7 @@ data Issue = Issue
   -- ^ Never optional. No 'Maybe'.
   , issueParent :: Maybe IssueId
   , issueProject :: Maybe (Id Project)
-  , issuePosition :: Int
+  , issuePosition :: Double
   , issueStage :: Maybe Int
   , issueStartDate :: Maybe CalendarDate
   , issueDueDate :: Maybe CalendarDate
